+++ dl_vsix/cli.py
@@ -1,9 +1,30 @@
-import argparse
 import json
+import os
 import shutil
 from pathlib import Path
 
+import typer
+
 from dl_vsix.dl import Extension, download_extensions
+from dl_vsix.extension_cache import DEFAULT_CACHE_MAXSIZE_MB, ExtensionCache
+
+# Initialize package cache in the global scope so we have access to a single instance for the CLI
+_cache_path_env = os.getenv("DL_VSIX_CACHE_DIR")
+_cache_path_override: Path | None
+if _cache_path_env is not None:
+    _cache_path_override = Path(_cache_path_env)
+else:
+    _cache_path_override = None
+
+_cache_maxsize_env = os.getenv("DL_VSIX_CACHE_DIR")
+if _cache_maxsize_env is not None:
+    _cache_maxsize_override = int(_cache_maxsize_env)
+else:
+    _cache_maxsize_override = DEFAULT_CACHE_MAXSIZE_MB
+
+PACKAGE_CACHE = ExtensionCache(
+    path_override=_cache_path_override, cache_maxsize_mb=_cache_maxsize_override
+)
 
 
 def _parse_extensions(spec_json: Path) -> list[Extension]:
@@ -30,39 +51,82 @@ def _parse_extensions(spec_json: Path) -> list[Extension]:
     return [Extension.from_id(s) for s in spec.get("extensions", [])]
 
 
-def main() -> None:  # noqa: D103
-    parser = argparse.ArgumentParser(
-        "dl_VSIX", description="Download VSIX bundles for offline extension installation."
-    )
-    source_group = parser.add_mutually_exclusive_group()
-    source_group.add_argument("extension", nargs="?", type=str, help="Single extension by ID")
-    source_group.add_argument(
-        "-s", "--spec_file", type=Path, help="JSON-specified collection of extensions"
-    )
-    parser.add_argument(
-        "-o",
-        "--out_dir",
-        type=Path,
-        default=Path("./vsix"),
-        help="Download directory (default: ./vsix)",
-    )
-    parser.add_argument(
-        "-f", "--follow_deps", action="store_false", help="Trace extension's dependencies."
-    )
-    parser.add_argument("-z", "--zip", action="store_true", help="Zip the download extension(s)")
-    args = parser.parse_args()
+dl_vsix_cli = typer.Typer(
+    no_args_is_help=True,
+    add_completion=False,
+    help="Download VSIX bundles for offline extension installation.",
+)
 
-    # Since we have a mutually exclusive group, only one of these can be specified
-    if args.extension is not None:
-        extensions = [Extension.from_id(args.extension)]
+
+@dl_vsix_cli.command(no_args_is_help=True)
+def download(
+    extension_id: str = typer.Argument("", help="Single extension by ID"),
+    spec_file: Path = typer.Option(
+        None, "-s", "--spec_file", help="JSON-specified collection of extensions", dir_okay=False
+    ),
+    out_dir: Path = typer.Option(
+        Path("./vsix"), "--out_dir", "-o", help="Download directory", file_okay=False
+    ),
+    follow_deps: bool = typer.Option(
+        False, "--follow_deps", "-f", help="Trace extension's dependencies"
+    ),
+    zip_result: bool = typer.Option(False, "--zip", "-z", help="Zip the download extension(s)"),
+) -> None:
+    """
+    Download VSIX extension packages.
+
+    NOTE: `extension_id` and `spec_file` are mutually exclusive.
+    """
+    if (not extension_id) and (spec_file is None):
+        raise ValueError("Either extension_id or spec_file must be specified.")
+
+    if extension_id and spec_file:
+        raise ValueError("Cannot specify both an extension_id and spec file.")
+
+    # Options are mutually exclusive, which should be enforced at this point
+    if extension_id:
+        extensions = [Extension.from_id(extension_id)]
     else:
-        extensions = _parse_extensions(args.spec_file)
+        extensions = _parse_extensions(spec_file)
 
-    download_extensions(extensions, out_dir=args.out_dir, follow_dependencies=args.follow_deps)
-    if args.zip:
-        zip_filepath = args.out_dir.parent / "zipped_extensions"
-        shutil.make_archive(base_name=zip_filepath, format="zip", root_dir=args.out_dir)
+    download_extensions(
+        extensions, out_dir=out_dir, package_cache=PACKAGE_CACHE, follow_dependencies=follow_deps
+    )
+    if zip_result:
+        zip_filepath = out_dir.parent / "zipped_extensions"
+        shutil.make_archive(base_name=str(zip_filepath), format="zip", root_dir=out_dir)
+
+
+cache_sub = typer.Typer(
+    name="cache", no_args_is_help=True, add_completion=False, help="Package cache utilities"
+)
+dl_vsix_cli.add_typer(cache_sub)
+
+
+@cache_sub.command("info")
+def cache_info() -> None:
+    """Show cache information."""
+    PACKAGE_CACHE.info()
+
+
+@cache_sub.command("list")
+def cache_list() -> None:
+    """List cache contents."""
+    PACKAGE_CACHE.list()
+
+
+@cache_sub.command("remove")
+def cache_remove(extensions: list[str]) -> None:
+    """Remove extension(s) from cache."""
+    for e in extensions:
+        PACKAGE_CACHE.remove(Extension.from_id(e))
+
+
+@cache_sub.command("purge")
+def cache_purge() -> None:
+    """Clear package cache."""
+    PACKAGE_CACHE.purge()
 
 
 if __name__ == "__main__":
-    main()
+    dl_vsix_cli()