+++ dl_vsix/dl.py
@@ -8,12 +8,20 @@ from pathlib import Path
 
 import httpx
 
+from dl_vsix.extension_cache import ExtensionCache
+from dl_vsix.extension_query import query_latest_version
+
 
 class Extension(t.NamedTuple):  # noqa: D101
     publisher: str
     extension: str
 
     def __str__(self) -> str:
+        return self.pID
+
+    @property
+    def pID(self) -> str:
+        """Build the full extension ID, as `'<publisher>.<package>'`."""
         return f"{self.publisher}.{self.extension}"
 
     @classmethod
@@ -27,13 +35,12 @@ class Extension(t.NamedTuple):  # noqa: D101
         publisher, extension = extension_id.split(".")
         return cls(publisher=publisher, extension=extension)
 
-    @property
-    def vsix_query(self) -> str:
+    def vsix_query(self, version: str = "latest") -> str:
         """Build query URL for the extension's latest VSIX package."""
         api_base = f"https://{self.publisher}.gallery.vsassets.io/_apis/public/gallery"
         publisher_comp = f"publisher/{self.publisher}"
         extension_comp = f"extension/{self.extension}"
-        suffix = "latest/assetbyname/Microsoft.VisualStudio.Services.VSIXPackage"
+        suffix = f"{version}/assetbyname/Microsoft.VisualStudio.Services.VSIXPackage"
 
         return f"{api_base}/{publisher_comp}/{extension_comp}/{suffix}"
 
@@ -65,10 +72,14 @@ def extract_dependencies(vsix_zip: Path, target: str = "extension/package.json")
 def download_extensions(
     extensions: list[Extension],
     out_dir: Path,
+    package_cache: ExtensionCache,
     follow_dependencies: bool = True,
 ) -> None:
     """
     Download VSIX packages for the specified extension(s) from the VS marketplace Gallery API.
+
+    Download cache management is provided using an instance of `ExtensionCache` passed to
+    `package_cache`.
 
     If `follow_dependencies` is `True`, the extension's metadata will be checked to see if it
     depends on any additional packages, which will be added to the queue if they haven't yet been
@@ -83,19 +94,29 @@ def download_extensions(
     with httpx.Client() as client:
         while extensions:
             ext = extensions.pop()
-            out_filepath = out_dir / f"{ext}.vsix"
-            with client.stream("GET", ext.vsix_query) as r:
-                if r.status_code == httpx.codes.OK:
-                    with out_filepath.open("wb") as f:
-                        for chunk in r.iter_bytes():
-                            f.write(chunk)
 
-                    seen_extensions.add(ext)
-                    print(f"Successfully downloaded extension '{ext}'")
+            latest_ver = query_latest_version(str(ext))
+            out_filepath = out_dir / f"{ext}_{latest_ver}.vsix"
 
-                else:
-                    print(f"Could not download extension '{ext}': {r.status_code}")
-                    continue
+            cached_ver = package_cache.cached_version(ext)
+            if (cached_ver is not None) and (cached_ver == latest_ver):
+                print(f"Cached download for '{ext}' found - version: {cached_ver}")
+                package_cache.copy_to(ext, out_dir)
+            else:
+                with client.stream("GET", ext.vsix_query(version=latest_ver)) as r:
+                    if r.status_code == httpx.codes.OK:
+                        with out_filepath.open("wb") as f:
+                            for chunk in r.iter_bytes():
+                                f.write(chunk)
+
+                        seen_extensions.add(ext)
+                        print(f"Successfully downloaded extension '{ext}', version: {latest_ver}")
+
+                        # Add to cache
+                        package_cache.insert(out_filepath)
+                    else:
+                        print(f"Could not download extension '{ext}': {r.status_code}")
+                        continue
 
             if follow_dependencies:
                 dependencies = extract_dependencies(out_filepath)